@@ -3,29 +3,18 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "@types/jest": "^24.0.0",
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.9.0",
-    "@types/react-router-dom": "^5.1.6",
-    "@types/react-router-redux": "^5.0.18",
-    "autobind-decorator": "^2.4.0",
-    "bootstrap": "^5.2.3",
-    "react": "^17.0.1",
-    "react-bootstrap": "^2.7.0",
-    "react-dom": "^17.0.1",
-    "react-redux": "^7.2.1",
-    "react-router-dom": "^5.2.0",
+    "bootstrap": "^5.3.1",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.8.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.1.2",
+    "react-router-dom": "^6.14.2",
     "react-scripts": "3.4.4",
     "set": "https://github.com/GeneralSet/Set",
-    "typescript": "~3.8.2",
-    "typestyle": "^2.1.0"
+    "typescript": "~5.1.6"
   },
   "scripts": {
-    "start": "react-app-rewired start",
+    "start": "export NODE_OPTIONS=--openssl-legacy-provider && react-app-rewired start",
     "build": "react-app-rewired build",
     "test": "react-app-rewired test",
     "eject": "react-app-rewired eject"
@@ -46,9 +35,17 @@
     ]
   },
   "devDependencies": {
-    "@types/react-redux": "^7.1.9",
-    "customize-cra": "^1.0.0",
-    "react-app-rewired": "^2.1.6",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.5.3",
+    "@types/node": "^20.4.8",
+    "@types/react": "^18.2.18",
+    "@types/react-dom": "^18.2.7",
+    "@types/react-redux": "^7.1.25",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-router-redux": "^5.0.22",
+    "react-app-rewired": "^2.2.1",
     "wasm-loader": "^1.3.0"
   }
 }
